pancake-order-system/workflow_worker/requirements.txt
annotated-types==0.7.0
protobuf==6.31.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
temporalio==1.11.1
types-protobuf==6.30.2.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
langchain>=0.3.25
inflect
redis
